Overwrite files when extracting with GNU tar
When using GNU tar, if a file already exists, from a previous run or
another cached action, it results in a slew of warning messages that
can cause issues for the run.

Non-GNU versions appear to automatically overwrite existing files.
From the BSD tar manpage:

```
-x    Extract to disk from the archive.  If a file with the same name
      appears more than once in the archive, each copy will be
      extracted, with later copies overwriting (replacing) earlier
      copies.
```

This change adds the use of `--overwrite` when using GNU tar so that
any existing files will just be overwritten as well.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>