Add another likely reason for problem matcher patterns failing
Input lines are passed as-is to the regular expressions used by problem matcher patterns. This can lead to unexpected failures when trying to, for example, match colored terminal output, which relies on [ANSI escape sequences](https://en.wikipedia.org/wiki/ANSI_escape_code). The cause is not that obvious at a first glance because such characters are stripped from the GitHub log UI, and I think that it should be documented.