[cache] Use Azure SDK to restore cache always if explicitly set to true
With the new cache service v2, `restoreCacheV2` [^1] and `saveCacheV2`
[^2] now explicitly use the Azure SDK for blob operations.

The cacheHttpClient uses the Azure SDK to save the cache if the
`useAzureSdk` option is set to true. [^3]

However, it may not use the Azure SDK to download the cache even if the
`useAzureSdk` option is set to true. If the signed URL does not end with
"blob.core.windows.net," it will not use the Azure SDK to restore the
cache. [^4]

To ensure consistent behavior between saving and restoring the cache,
the Azure SDK should be used to restore the cache if the `useAzureSdk`
option is explicitly set to true.

This PR ensures that the Azure SDK is used to restore the cache whenever
the `useAzureSdk` option is set to true.

[^1]: https://github.com/actions/toolkit/blob/1f7c2c79e034fe8a0d28006f52fc5b70f6dbb750/packages/cache/src/cache.ts#L224
[^2]: https://github.com/actions/toolkit/blob/1f7c2c79e034fe8a0d28006f52fc5b70f6dbb750/packages/cache/src/cache.ts#L472
[^3]: https://github.com/actions/toolkit/blob/1f7c2c79e034fe8a0d28006f52fc5b70f6dbb750/packages/cache/src/internal/cacheHttpClient.ts#L335-L343
[^4]: https://github.com/actions/toolkit/blob/1f7c2c79e034fe8a0d28006f52fc5b70f6dbb750/packages/cache/src/internal/cacheHttpClient.ts#L153